requirements39_Windows.txt
backcall==0.2.0
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.0
decorator==5.1.0
EDFlib-Python==1.0.6
eeglabio==0.0.2.post4
entrypoints==0.3
ipykernel==6.6.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.1.0
jupyter-core==4.8.1
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
mne==1.5.1
nest-asyncio==1.5.1
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
Pygments==2.10.0
pylsl==1.15.0
pyparsing==2.4.7
pyqtgraph==0.12.3
pyserial==3.5
PySide2==5.15.2
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pyxdf==1.16.3
pyzmq==22.3.0
scipy==1.7.1
shiboken2==5.15.2
six==1.16.0
spyder-kernels==2.2.1
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5